Struct textplots::Chart

pub struct Chart<'a> { /* private fields */ }

Controls the drawing.

Implementations

impl<'a> Chart<'a>

pub fn new(width: u32, height: u32, xmin: f32, xmax: f32) -> Self

Creates a new Chart object.

Panics

Panics if width is less than 32 or height is less than 3.

pub fn new_with_y_range( width: u32, height: u32, xmin: f32, xmax: f32, ymin: f32, ymax: f32 ) -> Self

Creates a new Chart object with fixed y axis range.

Panics

Panics if width is less than 32 or height is less than 3.

pub fn display(&mut self)

Prints canvas content.

pub fn nice(&mut self)

Prints canvas content with some additional visual elements (like borders).

pub fn axis(&mut self)

Shows axis.

pub fn x_axis(&mut self)

Shows x-axis.

pub fn y_axis(&mut self)

Shows y-axis.

pub fn figures(&mut self)

pub fn frame(&self) -> String

Returns the frame.